refactor: refactor HTTP client abstraction and add `surf` support
