+++ crates/tombi-schema-store/src/http_client/reqwest_client.rs
@@ -1,0 +1,35 @@
+use crate::http_client::error::FetchError;
+use bytes::Bytes;
+
+#[derive(Debug, Clone)]
+pub struct HttpClient(reqwest::Client);
+
+impl HttpClient {
+    pub fn new() -> Self {
+        Self(reqwest::Client::new())
+    }
+
+    pub async fn get_bytes(&self, url: &str) -> Result<Bytes, FetchError> {
+        let response = self
+            .0
+            .get(url)
+            .send()
+            .await
+            .map_err(|err| FetchError::FetchFailed {
+                reason: err.to_string(),
+            })?;
+
+        if !response.status().is_success() {
+            return Err(FetchError::StatusNotOk {
+                status: response.status().as_u16(),
+            });
+        }
+
+        response
+            .bytes()
+            .await
+            .map_err(|err| FetchError::BodyReadFailed {
+                reason: err.to_string(),
+            })
+    }
+}
